+++ changelog.md
@@ -1,3 +1,8 @@
+# [v1.1.0] 2022-02-08
+
+- [MINOR] Add --init cli option to initialize config
+- [PATCH] Ignore config when --sort-only option is used
+
 # [v1.0.4] 2023-02-07
 
 - [PATCH] Fix CRLF issue on Mac/Linux